+++ mycodo/inputs/atlas_humidity.py
@@ -170,7 +170,7 @@ class InputModule(AbstractInput):
                     break
 
         elif self.interface == 'I2C':
-            return_string = atlas_return
+            return_string = self.atlas_device.build_string(atlas_return)
 
         # Parse return string
         if return_string and ',' in return_string: